requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
psycopg2-binary
alembic
pydantic
pydantic-settings
requests
faster-whisper
python-multipart
google-generativeai
pydub
geopy
google-genai
pgvector
httpx
python-jose[cryptography]
bcrypt
pydantic[email]
aiohttp
jose